Moved MainActivity into new package
