+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             tools:ignore="GoogleAppIndexingWarning"
             tools:targetApi="m">
 
-        <activity android:name=".MainActivity"
+        <activity android:name=".activities.MainActivity"
                   android:launchMode="singleTop">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
